requirements.txt
flask>=3.0
pikepdf>=8.0
Pillow>=10.0
reportlab>=4.0
